New design
It's about time I update my personal website. The design is 7 years old
and the content 4 years old. Let's start refreshing it.

This introduces a new, simpler design, removing all the existing CSS and
replacing it with a lightweight stylesheet to begin with. I will make
sure I add to it sparingly, to keep its size as low as possible.