Redesign website header
Updates the styles so that the website header looks nicer, with all
elements on the same line and the navigation aligned to the right.

Removes the `page` layout and updates all pages to use the `default`
one, which now shows the page's title if present in the front matter.

Finally, updates the `title` tag in the `head` section, tweaks the
various `link` tags and cleans-up the HTML markup.