Switch from Jekyll to Next.js
Since I moved to Vercel (formerly Zeit) to host my website, it makes
sence to use Next.js, which the platform has been built around. I also
want to improve my JavaScript skills, so this seems like a good move.

Ports the website and its content to Next.js as it is.