Rename blog "posts" to "thoughts"
I'm not sure if this is a great idea, but I like the sound of "reads"
and "thoughts", so let's give it a try.

This does all the necessary renaming for the transition, so it should be
fairly seemless. It also moves the Markdown files to `data/posts` and
the images to `public/images/posts` to keep things tidy.